fix(web): touchpath recordings always sample end of path, even when stationary 🐵

+++ common/web/gesture-recognizer/src/engine/inputEventEngine.ts
@@ -89,9 +89,27 @@ export abstract class InputEventEngine<HoveredItemType> extends EventEmitter<Eve
     touchpoint.path.terminate(true);
   }
 
-  protected onInputEnd(identifier: number) {
-    // We do not add extend the path here because any 'end' event immediately
-    // follows a 'move' if it occurred simultaneously.
+  protected onInputEnd(identifier: number, target: EventTarget) {
+    const touchpoint = this.getTouchpointWithId(identifier);
+    if(!touchpoint) {
+      return;
+    }
+
+    const lastEntry = touchpoint.path.coords[touchpoint.path.coords.length-1];
+    const sample = this.buildSampleFor(lastEntry.clientX, lastEntry.clientY, target);
+
+    /* While an 'end' event immediately follows a 'move' if it occurred simultaneously,
+     * this is decidedly _not_ the case if the touchpoint was held for a while without
+     * moving, even at the point of its release.
+     *
+     * We'll never need to worry about the touchpoint moving here, and thus we don't
+     * need to worry about `currentHoveredItem` changing.  We're only concerned with
+     * recording the _timing_ of the touchpoint's release.
+     */
+    if(sample.t != lastEntry.t) {
+      touchpoint.update(sample);
+    }
+
     this.getTouchpointWithId(identifier)?.path.terminate(false);
   }
 }
